Patch by Wim van der Meer: Added get_mouse_bitmap() global method to InterfaceDefs.h and corresponding implementation in the app_server. Thanks a bunch! Closes ticket #5978.
A note to those using "update-all" with hybrid builds - the alternative GCC
system libs will not be updated this way, and this and Wim's last patch
change the libbe <-> app_server interface. You need to manually update the
respective libbe.so, or do a clean build, otherwise apps for the other GCC will
not start anymore.


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@36830 a95241bf-73f2-0310-859d-f6bbb57e9c96